backend/user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
python-jose[cryptography]    # For JWT
slowapi                      # For rate limiting
unidecode
pydantic[email]
requests